requirements.txt
absl-py==2.3.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
editdistance==0.8.1
einops==0.8.1
emotiefflib==1.0
executing==2.2.0
facenet-pytorch==2.6.0
fer==22.5.1
ffmpeg==1.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
funasr==1.0.27
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kaldiio==2.18.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
modelscope==1.24.1
more-itertools==10.6.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
ollama==0.4.7
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
oss2==2.19.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.7
pooch==1.8.2
proglog==0.1.12
prompt_toolkit==3.0.50
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.1
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rotary-embedding-torch==0.8.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
snakeviz==2.2.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.2.2
torch-complex==0.4.4
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
triton==3.2.0
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
